[Bronze V] Title: 이상한 기호, Time: 204 ms, Memory: 17692 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -2,8 +2,8 @@ import java.util.Scanner;
 public class Main {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
-		int num1 = sc.nextInt();
-		int num2 = sc.nextInt();
+		long num1 = sc.nextLong();
+		long num2 = sc.nextLong();
 		
 		System.out.println((num1+num2)*(num1-num2));
 	}
